+++ src/Xreferee/Lsp.hs
@@ -80,9 +80,7 @@ run cliOptions = flip E.catches handlers do
       -- Log everything to a file if the user specified a log file path, otherwise do nothing.
       fileLogger :: LogAction IO (WithSeverity Text)
       fileLogger =
-        maybeLogFileHandle
-          <&> (\logFileHandle -> LogAction $ \msg -> T.hPutStrLn logFileHandle (getMsg msg))
-          & fromMaybe mempty
+        maybe mempty (\logFileHandle -> LogAction $ \msg -> T.hPutStrLn logFileHandle (getMsg msg)) maybeLogFileHandle
 
       -- During startup, before we have a connection to the client:
       --   * Log everything to stderr
@@ -222,8 +220,7 @@ handlers :: Handlers AppM
 handlers =
   mconcat
     [ notificationHandler LSP.SMethod_Initialized $ Util.timedNot \_msg -> do
-        FileWatchers.watchRepoFiles
-        modifyState $ sendDiagnostics,
+        FileWatchers.watchRepoFiles,
       notificationHandler LSP.SMethod_TextDocumentDidOpen (Util.timedNot $ filterNot handleDidOpen),
       notificationHandler LSP.SMethod_TextDocumentDidClose \_req -> do
         -- Empty handler so we don't get these warnings in the log: `LSP: no handler for: "textDocument/didClose"`
@@ -249,7 +246,7 @@ handlers =
       Handler AppM method
     filterReq handler = \msg responder -> do
       let uri = msg ^. LSP.params . LSP.textDocument . LSP.uri
-      whenM (Util.shouldHandleFile uri) do
+      whenM (Util.shouldHandleFileOrDir uri) do
         handler msg responder
 
     -- Skip the handler if we're not interested in processing events for this file
@@ -261,7 +258,7 @@ handlers =
       Handler AppM method
     filterNot handler = \msg -> do
       let uri = msg ^. LSP.params . LSP.textDocument . LSP.uri
-      whenM (Util.shouldHandleFile uri) do
+      whenM (Util.shouldHandleFileOrDir uri) do
         handler msg
 
 -- | Dump an exception, along with a timestamp and some context, to a crash log file on disk,