+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 * Dump to a logfile when the server crashes.
 * Display runtime exceptions in the client.
+* Bug fixes:
+  * Scan tracked & git-ignored files, see: https://github.com/brandonchinn178/xreferee/issues/27
+  * Treat paths as literal pathspecs when checking whether a file is binary via `git ls-files`
+  * Fixed race conditions when files were deleted after an event is triggered and before it's done being handled.
+  * Don't follow symlinks (to match `xreferee`'s behavior)
+  * When a folder is created, individually check whether each file should be handled or ignored.
 
 ## 0.0.2